requirements.txt
streamlit==1.38.0
gspread==6.1.2
google-auth==2.34.0
streamlit-authenticator==0.3.2
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
tzdata==2024.1
pytest==8.3.2
black==24.8.0
